create test dir
